Get reference list for publications from crossref and opencitations

Get reference list for publications from crossref and opencitations

Usage

reference(remove_old = TRUE, all = FALSE)

Arguments

remove_old

whether to remove old files

all

whether to get all reference

Value

no return

Details

The reference list is obtained from crossref and opencitations. The field `reference` in Tiddlywiki is updated to list all reference in the Tiddlywiki. Two fields `reference-count` and `cited-count` is also updated for number of reference and citation for a publication, respectively.